@@ -35,10 +35,13 @@
 #include "pilz_industrial_motion_planner/trajectory_blender_transition_window.h"
 
 #include <algorithm>
+#include <memory>
 #include <math.h>
 #include <tf2_eigen/tf2_eigen.h>
+#include <moveit/planning_interface/planning_interface.h>
 
 bool pilz_industrial_motion_planner::TrajectoryBlenderTransitionWindow::blend(
+    const planning_scene::PlanningSceneConstPtr& planning_scene,
     const pilz_industrial_motion_planner::TrajectoryBlendRequest& req,
     pilz_industrial_motion_planner::TrajectoryBlendResponse& res)
 {
@@ -85,10 +88,10 @@ bool pilz_industrial_motion_planner::TrajectoryBlenderTransitionWindow::blend(
   }
   trajectory_msgs::JointTrajectory blend_joint_trajectory;
   moveit_msgs::MoveItErrorCodes error_code;
-  if (!generateJointTrajectory(req.first_trajectory->getFirstWayPointPtr()->getRobotModel(),
-                               limits_.getJointLimitContainer(), blend_trajectory_cartesian, req.group_name,
-                               req.link_name, initial_joint_position, initial_joint_velocity, blend_joint_trajectory,
-                               error_code, true))
+
+  if (!generateJointTrajectory(planning_scene, limits_.getJointLimitContainer(), blend_trajectory_cartesian,
+                               req.group_name, req.link_name, initial_joint_position, initial_joint_velocity,
+                               blend_joint_trajectory, error_code, true))
   {
     // LCOV_EXCL_START
     ROS_INFO("Failed to generate joint trajectory for blending trajectory.");
@@ -97,12 +100,12 @@ bool pilz_industrial_motion_planner::TrajectoryBlenderTransitionWindow::blend(
     // LCOV_EXCL_STOP
   }
 
-  res.first_trajectory = std::shared_ptr<robot_trajectory::RobotTrajectory>(
-      new robot_trajectory::RobotTrajectory(req.first_trajectory->getRobotModel(), req.first_trajectory->getGroup()));
-  res.blend_trajectory = std::shared_ptr<robot_trajectory::RobotTrajectory>(
-      new robot_trajectory::RobotTrajectory(req.first_trajectory->getRobotModel(), req.first_trajectory->getGroup()));
-  res.second_trajectory = std::shared_ptr<robot_trajectory::RobotTrajectory>(
-      new robot_trajectory::RobotTrajectory(req.first_trajectory->getRobotModel(), req.first_trajectory->getGroup()));
+  res.first_trajectory = std::make_shared<robot_trajectory::RobotTrajectory>(req.first_trajectory->getRobotModel(),
+                                                                             req.first_trajectory->getGroup());
+  res.blend_trajectory = std::make_shared<robot_trajectory::RobotTrajectory>(req.first_trajectory->getRobotModel(),
+                                                                             req.first_trajectory->getGroup());
+  res.second_trajectory = std::make_shared<robot_trajectory::RobotTrajectory>(req.first_trajectory->getRobotModel(),
+                                                                              req.first_trajectory->getGroup());
 
   // set the three trajectories after blending in response
   // erase the points [first_intersection_index, back()] from the first
