@@ -74,8 +74,8 @@ protected:
     // Load the plugin
     try
     {
-      planner_plugin_loader_.reset(new pluginlib::ClassLoader<planning_interface::PlannerManager>(
-          "moveit_core", "planning_interface::PlannerManager"));
+      planner_plugin_loader_ = std::make_unique<pluginlib::ClassLoader<planning_interface::PlannerManager>>(
+          "moveit_core", "planning_interface::PlannerManager");
     }
     catch (pluginlib::PluginlibException& ex)
     {
@@ -203,6 +203,7 @@ TEST_P(CommandPlannerTest, CheckPlanningContextRequest)
   moveit_msgs::MotionPlanRequest req;
   moveit_msgs::MoveItErrorCodes error_code;
 
+  req.group_name = "manipulator";
   // Check for the algorithms
   std::vector<std::string> algs;
   planner_instance_->getPlanningAlgorithms(algs);
