@@ -82,6 +82,7 @@ protected:
   // ros stuff
   ros::NodeHandle ph_{ "~" };
   robot_model::RobotModelConstPtr robot_model_{ robot_model_loader::RobotModelLoader(GetParam()).getModel() };
+  planning_scene::PlanningSceneConstPtr planning_scene_{ new planning_scene::PlanningScene(robot_model_) };
 
   // trajectory generator
   std::unique_ptr<TrajectoryGenerator> ptp_;
@@ -126,7 +127,7 @@ void TrajectoryGeneratorPTPTest::SetUp()
 
   // create the trajectory generator
   planner_limits_.setJointLimits(joint_limits);
-  ptp_.reset(new TrajectoryGeneratorPTP(robot_model_, planner_limits_));
+  ptp_ = std::make_unique<TrajectoryGeneratorPTP>(robot_model_, planner_limits_, planning_group_);
   ASSERT_NE(nullptr, ptp_);
 }
 
@@ -168,7 +169,8 @@ INSTANTIATE_TEST_SUITE_P(InstantiationName, TrajectoryGeneratorPTPTest,
 TEST_P(TrajectoryGeneratorPTPTest, noLimits)
 {
   LimitsContainer planner_limits;
-  EXPECT_THROW(TrajectoryGeneratorPTP(this->robot_model_, planner_limits), TrajectoryGeneratorInvalidLimitsException);
+  EXPECT_THROW(TrajectoryGeneratorPTP(this->robot_model_, planner_limits, planning_group_),
+               TrajectoryGeneratorInvalidLimitsException);
 }
 
 /**
@@ -194,7 +196,7 @@ TEST_P(TrajectoryGeneratorPTPTest, emptyRequest)
 
   EXPECT_FALSE(res.trajectory_->empty());
 
-  EXPECT_FALSE(ptp_->generate(req, res));
+  EXPECT_FALSE(ptp_->generate(planning_scene_, req, res));
 
   EXPECT_TRUE(res.trajectory_->empty());
 }
@@ -220,7 +222,8 @@ TEST_P(TrajectoryGeneratorPTPTest, missingVelocityLimits)
   }
 
   planner_limits.setJointLimits(joint_limits);
-  EXPECT_THROW(TrajectoryGeneratorPTP(this->robot_model_, planner_limits), TrajectoryGeneratorInvalidLimitsException);
+  EXPECT_THROW(TrajectoryGeneratorPTP(this->robot_model_, planner_limits, planning_group_),
+               TrajectoryGeneratorInvalidLimitsException);
 }
 
 /**
@@ -243,7 +246,8 @@ TEST_P(TrajectoryGeneratorPTPTest, missingDecelerationimits)
   }
 
   planner_limits.setJointLimits(joint_limits);
-  EXPECT_THROW(TrajectoryGeneratorPTP(this->robot_model_, planner_limits), TrajectoryGeneratorInvalidLimitsException);
+  EXPECT_THROW(TrajectoryGeneratorPTP(this->robot_model_, planner_limits, planning_group_),
+               TrajectoryGeneratorInvalidLimitsException);
 }
 
 /**
@@ -286,7 +290,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testInsufficientLimit)
   EXPECT_THROW(
       {
         std::unique_ptr<TrajectoryGeneratorPTP> ptp_error(
-            new TrajectoryGeneratorPTP(robot_model_, insufficient_planner_limits));
+            new TrajectoryGeneratorPTP(robot_model_, insufficient_planner_limits, planning_group_));
       },
       TrajectoryGeneratorInvalidLimitsException);
 
@@ -325,7 +329,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testInsufficientLimit)
 
   EXPECT_NO_THROW({
     std::unique_ptr<TrajectoryGeneratorPTP> ptp_no_error(
-        new TrajectoryGeneratorPTP(robot_model_, sufficient_planner_limits));
+        new TrajectoryGeneratorPTP(robot_model_, sufficient_planner_limits, planning_group_));
   });
 }
 
@@ -359,7 +363,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testCartesianGoal)
   //****************************************
   //*** test robot model without gripper ***
   //****************************************
-  ASSERT_TRUE(ptp_->generate(req, res));
+  ASSERT_TRUE(ptp_->generate(planning_scene_, req, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::SUCCESS);
 
   moveit_msgs::MotionPlanResponse res_msg;
@@ -407,12 +411,12 @@ TEST_P(TrajectoryGeneratorPTPTest, testCartesianGoalMissingLinkNameConstraints)
 
   planning_interface::MotionPlanRequest req_no_position_constaint_link_name = req;
   req_no_position_constaint_link_name.goal_constraints.front().position_constraints.front().link_name = "";
-  ASSERT_FALSE(ptp_->generate(req_no_position_constaint_link_name, res));
+  ASSERT_FALSE(ptp_->generate(planning_scene_, req_no_position_constaint_link_name, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 
   planning_interface::MotionPlanRequest req_no_orientation_constaint_link_name = req;
   req_no_orientation_constaint_link_name.goal_constraints.front().orientation_constraints.front().link_name = "";
-  ASSERT_FALSE(ptp_->generate(req_no_orientation_constaint_link_name, res));
+  ASSERT_FALSE(ptp_->generate(planning_scene_, req_no_orientation_constaint_link_name, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
@@ -439,7 +443,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testInvalidCartesianGoal)
       kinematic_constraints::constructGoalConstraints(target_link_, pose, tolerance_pose, tolerance_angle);
   req.goal_constraints.push_back(pose_goal);
 
-  ASSERT_FALSE(ptp_->generate(req, res));
+  ASSERT_FALSE(ptp_->generate(planning_scene_, req, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::NO_IK_SOLUTION);
   EXPECT_EQ(res.trajectory_, nullptr);
 }
@@ -465,7 +469,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testJointGoalAlreadyReached)
   req.goal_constraints.push_back(gc);
 
   // TODO lin and circ has different settings
-  ASSERT_TRUE(ptp_->generate(req, res));
+  ASSERT_TRUE(ptp_->generate(planning_scene_, req, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::SUCCESS);
 
   moveit_msgs::MotionPlanResponse res_msg;
@@ -516,7 +520,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testScalingFactor)
   planner_limits.setJointLimits(joint_limits);
 
   // create the generator with new limits
-  ptp_.reset(new TrajectoryGeneratorPTP(robot_model_, planner_limits));
+  ptp_ = std::make_unique<TrajectoryGeneratorPTP>(robot_model_, planner_limits, planning_group_);
 
   planning_interface::MotionPlanResponse res;
   planning_interface::MotionPlanRequest req;
@@ -537,7 +541,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testScalingFactor)
   req.max_velocity_scaling_factor = 0.5;
   req.max_acceleration_scaling_factor = 1.0 / 3.0;
 
-  ASSERT_TRUE(ptp_->generate(req, res));
+  ASSERT_TRUE(ptp_->generate(planning_scene_, req, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::SUCCESS);
 
   moveit_msgs::MotionPlanResponse res_msg;
@@ -664,7 +668,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testJointGoalAndAlmostZeroStartVelocity)
   gc.joint_constraints.push_back(jc);
   req.goal_constraints.push_back(gc);
 
-  ASSERT_TRUE(ptp_->generate(req, res));
+  ASSERT_TRUE(ptp_->generate(planning_scene_, req, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::SUCCESS);
 
   moveit_msgs::MotionPlanResponse res_msg;
@@ -806,7 +810,7 @@ TEST_P(TrajectoryGeneratorPTPTest, testJointGoalNoStartVel)
   gc.joint_constraints.push_back(jc);
   req.goal_constraints.push_back(gc);
 
-  ASSERT_TRUE(ptp_->generate(req, res));
+  ASSERT_TRUE(ptp_->generate(planning_scene_, req, res));
   EXPECT_EQ(res.error_code_.val, moveit_msgs::MoveItErrorCodes::SUCCESS);
 
   moveit_msgs::MotionPlanResponse res_msg;
