@@ -49,7 +49,7 @@
 
 namespace pilz_industrial_motion_planner
 {
-MOVEIT_CLASS_FORWARD(PlanningContext)
+MOVEIT_CLASS_FORWARD(PlanningContext);
 
 /**
  * @brief PlanningContext for obtaining CIRC trajectories
