@@ -41,6 +41,7 @@
 #include <moveit/robot_trajectory/robot_trajectory.h>
 #include <tf2/transform_datatypes.h>
 #include <trajectory_msgs/MultiDOFJointTrajectory.h>
+#include <moveit/planning_scene/planning_scene.h>
 
 #include "pilz_industrial_motion_planner/cartesian_trajectory.h"
 #include "pilz_industrial_motion_planner/limits_container.h"
@@ -50,6 +51,7 @@ namespace pilz_industrial_motion_planner
 /**
  * @brief compute the inverse kinematics of a given pose, also check robot self
  * collision
+ * @param scene: planning scene
  * @param robot_model: kinematic model of the robot
  * @param group_name: name of planning group
  * @param link_name: name of target link
@@ -62,12 +64,12 @@ namespace pilz_industrial_motion_planner
  * @param timeout: timeout for IK, if not set the default solver timeout is used
  * @return true if succeed
  */
-bool computePoseIK(const robot_model::RobotModelConstPtr& robot_model, const std::string& group_name,
+bool computePoseIK(const planning_scene::PlanningSceneConstPtr& scene, const std::string& group_name,
                    const std::string& link_name, const Eigen::Isometry3d& pose, const std::string& frame_id,
                    const std::map<std::string, double>& seed, std::map<std::string, double>& solution,
                    bool check_self_collision = true, const double timeout = 0.0);
 
-bool computePoseIK(const robot_model::RobotModelConstPtr& robot_model, const std::string& group_name,
+bool computePoseIK(const planning_scene::PlanningSceneConstPtr& scene, const std::string& group_name,
                    const std::string& link_name, const geometry_msgs::Pose& pose, const std::string& frame_id,
                    const std::map<std::string, double>& seed, std::map<std::string, double>& solution,
                    bool check_self_collision = true, const double timeout = 0.0);
@@ -107,6 +109,7 @@ bool verifySampleJointLimits(const std::map<std::string, double>& position_last,
 
 /**
  * @brief Generate joint trajectory from a KDL Cartesian trajectory
+ * @param scene: planning scene
  * @param robot_model: robot kinematics model
  * @param joint_limits: joint limits
  * @param trajectory: KDL Cartesian trajectory
@@ -122,7 +125,7 @@ bool verifySampleJointLimits(const std::map<std::string, double>& position_last,
  * @param check_self_collision: check for self collision during creation
  * @return true if succeed
  */
-bool generateJointTrajectory(const robot_model::RobotModelConstPtr& robot_model,
+bool generateJointTrajectory(const planning_scene::PlanningSceneConstPtr& scene,
                              const JointLimitsContainer& joint_limits, const KDL::Trajectory& trajectory,
                              const std::string& group_name, const std::string& link_name,
                              const std::map<std::string, double>& initial_joint_position, const double& sampling_time,
@@ -131,6 +134,7 @@ bool generateJointTrajectory(const robot_model::RobotModelConstPtr& robot_model,
 
 /**
  * @brief Generate joint trajectory from a MultiDOFJointTrajectory
+ * @param scene: planning scene
  * @param trajectory: Cartesian trajectory
  * @param info: motion plan information
  * @param sampling_time
@@ -138,7 +142,7 @@ bool generateJointTrajectory(const robot_model::RobotModelConstPtr& robot_model,
  * @param error_code
  * @return true if succeed
  */
-bool generateJointTrajectory(const robot_model::RobotModelConstPtr& robot_model,
+bool generateJointTrajectory(const planning_scene::PlanningSceneConstPtr& scene,
                              const JointLimitsContainer& joint_limits,
                              const pilz_industrial_motion_planner::CartesianTrajectory& trajectory,
                              const std::string& group_name, const std::string& link_name,
@@ -202,6 +206,7 @@ bool intersectionFound(const Eigen::Vector3d& p_center, const Eigen::Vector3d& p
 
 /**
  * @brief Checks if current robot state is in self collision.
+ * @param scene: planning scene.
  * @param test_for_self_collision Flag to deactivate this check during IK.
  * @param robot_model: robot kinematics model.
  * @param state Robot state instance used for .
@@ -209,9 +214,26 @@ bool intersectionFound(const Eigen::Vector3d& p_center, const Eigen::Vector3d& p
  * @param ik_solution
  * @return
  */
-bool isStateColliding(const bool test_for_self_collision, const moveit::core::RobotModelConstPtr& robot_model,
+bool isStateColliding(const bool test_for_self_collision, const planning_scene::PlanningSceneConstPtr& scene,
                       robot_state::RobotState* state, const robot_state::JointModelGroup* const group,
                       const double* const ik_solution);
 }  // namespace pilz_industrial_motion_planner
 
 void normalizeQuaternion(geometry_msgs::Quaternion& quat);
+
+/**
+ * @brief Adapt goal pose, defined by position+orientation, to consider offset
+ * @param constraint to apply offset to
+ * @param offset to apply to the constraint
+ * @param orientation to apply to the offset
+ * @return final goal pose
+ */
+Eigen::Isometry3d getConstraintPose(const geometry_msgs::Point& position, const geometry_msgs::Quaternion& orientation,
+                                    const geometry_msgs::Vector3& offset);
+
+/**
+ * @brief Conviencency method, passing args from a goal constraint
+ * @param goal goal constraint
+ * @return final goal pose
+ */
+Eigen::Isometry3d getConstraintPose(const moveit_msgs::Constraints& goal);
