@@ -42,6 +42,7 @@
 #include <kdl/trajectory.hpp>
 #include <moveit/planning_interface/planning_interface.h>
 #include <moveit/robot_model/robot_model.h>
+#include <moveit/planning_scene/planning_scene.h>
 
 #include "pilz_industrial_motion_planner/joint_limits_extension.h"
 #include "pilz_industrial_motion_planner/limits_container.h"
@@ -103,10 +104,10 @@ public:
    * @param res: motion plan response
    * @param sampling_time: sampling time of the generate trajectory
    * @return motion plan succeed/fail, detailed information in motion plan
-   * responce
+   * response
    */
-  bool generate(const planning_interface::MotionPlanRequest& req, planning_interface::MotionPlanResponse& res,
-                double sampling_time = 0.1);
+  bool generate(const planning_scene::PlanningSceneConstPtr& scene, const planning_interface::MotionPlanRequest& req,
+                planning_interface::MotionPlanResponse& res, double sampling_time = 0.1);
 
 protected:
   /**
@@ -144,13 +145,16 @@ private:
    * @brief Extract needed information from a motion plan request in order to
    * simplify
    * further usages.
+   * @param scene: planning scene
    * @param req: motion plan request
    * @param info: information extracted from motion plan request which is
    * necessary for the planning
    */
-  virtual void extractMotionPlanInfo(const planning_interface::MotionPlanRequest& req, MotionPlanInfo& info) const = 0;
+  virtual void extractMotionPlanInfo(const planning_scene::PlanningSceneConstPtr& scene,
+                                     const planning_interface::MotionPlanRequest& req, MotionPlanInfo& info) const = 0;
 
-  virtual void plan(const planning_interface::MotionPlanRequest& req, const MotionPlanInfo& plan_info,
+  virtual void plan(const planning_scene::PlanningSceneConstPtr& scene,
+                    const planning_interface::MotionPlanRequest& req, const MotionPlanInfo& plan_info,
                     const double& sampling_time, trajectory_msgs::JointTrajectory& joint_trajectory) = 0;
 
 private:
@@ -200,7 +204,7 @@ private:
   /**
    * @brief set MotionPlanResponse from joint trajectory
    */
-  void setSuccessResponse(const std::string& group_name, const moveit_msgs::RobotState& start_state,
+  void setSuccessResponse(const moveit::core::RobotState& start_rs, const std::string& group_name,
                           const trajectory_msgs::JointTrajectory& joint_trajectory, const ros::Time& planning_start,
                           planning_interface::MotionPlanResponse& res) const;
 
@@ -229,10 +233,6 @@ private:
                                 const std::string& group_name) const;
 
   void checkCartesianGoalConstraint(const moveit_msgs::Constraints& constraint, const std::string& group_name) const;
-
-  void convertToRobotTrajectory(const trajectory_msgs::JointTrajectory& joint_trajectory,
-                                const moveit_msgs::RobotState& start_state,
-                                robot_trajectory::RobotTrajectory& robot_trajectory) const;
 
 private:
   /**
