@@ -34,6 +34,8 @@
 
 #pragma once
 
+#include <moveit/planning_scene/planning_scene.h>
+
 #include "eigen3/Eigen/Eigen"
 #include "pilz_industrial_motion_planner/trajectory_generation_exceptions.h"
 #include "pilz_industrial_motion_planner/trajectory_generator.h"
@@ -61,10 +63,12 @@ public:
    * @param model: a map of joint limits information
    */
   TrajectoryGeneratorPTP(const robot_model::RobotModelConstPtr& robot_model,
-                         const pilz_industrial_motion_planner::LimitsContainer& planner_limits);
+                         const pilz_industrial_motion_planner::LimitsContainer& planner_limits,
+                         const std::string& group_name);
 
 private:
-  void extractMotionPlanInfo(const planning_interface::MotionPlanRequest& req, MotionPlanInfo& info) const override;
+  void extractMotionPlanInfo(const planning_scene::PlanningSceneConstPtr& scene,
+                             const planning_interface::MotionPlanRequest& req, MotionPlanInfo& info) const override;
 
   /**
    * @brief plan ptp joint trajectory with zero start velocity
@@ -77,18 +81,18 @@ private:
    * @param sampling_time
    */
   void planPTP(const std::map<std::string, double>& start_pos, const std::map<std::string, double>& goal_pos,
-               trajectory_msgs::JointTrajectory& joint_trajectory, const std::string& group_name,
-               const double& velocity_scaling_factor, const double& acceleration_scaling_factor,
-               const double& sampling_time);
+               trajectory_msgs::JointTrajectory& joint_trajectory, const double& velocity_scaling_factor,
+               const double& acceleration_scaling_factor, const double& sampling_time);
 
-  void plan(const planning_interface::MotionPlanRequest& req, const MotionPlanInfo& plan_info,
-            const double& sampling_time, trajectory_msgs::JointTrajectory& joint_trajectory) override;
+  void plan(const planning_scene::PlanningSceneConstPtr& scene, const planning_interface::MotionPlanRequest& req,
+            const MotionPlanInfo& plan_info, const double& sampling_time,
+            trajectory_msgs::JointTrajectory& joint_trajectory) override;
 
 private:
   const double MIN_MOVEMENT = 0.001;
   pilz_industrial_motion_planner::JointLimitsContainer joint_limits_;
-  // most strict joint limits for each group
-  std::map<std::string, JointLimit> most_strict_limits_;
+  // most strict joint limits
+  JointLimit most_strict_limit_;
 };
 
 }  // namespace pilz_industrial_motion_planner
