@@ -68,7 +68,8 @@ void PlanComponentsBuilder::appendWithStrictTimeIncrease(robot_trajectory::Robot
   }
 }
 
-void PlanComponentsBuilder::blend(const robot_trajectory::RobotTrajectoryPtr& other, const double blend_radius)
+void PlanComponentsBuilder::blend(const planning_scene::PlanningSceneConstPtr& planning_scene,
+                                  const robot_trajectory::RobotTrajectoryPtr& other, const double blend_radius)
 {
   if (!blender_)
   {
@@ -86,7 +87,7 @@ void PlanComponentsBuilder::blend(const robot_trajectory::RobotTrajectoryPtr& ot
   blend_request.link_name = getSolverTipFrame(model_->getJointModelGroup(blend_request.group_name));
 
   pilz_industrial_motion_planner::TrajectoryBlendResponse blend_response;
-  if (!blender_->blend(blend_request, blend_response))
+  if (!blender_->blend(planning_scene, blend_request, blend_response))
   {
     throw BlendingFailedException("Blending failed");
   }
@@ -98,7 +99,8 @@ void PlanComponentsBuilder::blend(const robot_trajectory::RobotTrajectoryPtr& ot
   traj_tail_ = blend_response.second_trajectory;  // first for next blending segment
 }
 
-void PlanComponentsBuilder::append(const robot_trajectory::RobotTrajectoryPtr& other, const double blend_radius)
+void PlanComponentsBuilder::append(const planning_scene::PlanningSceneConstPtr& planning_scene,
+                                   const robot_trajectory::RobotTrajectoryPtr& other, const double blend_radius)
 {
   if (!model_)
   {
@@ -131,7 +133,7 @@ void PlanComponentsBuilder::append(const robot_trajectory::RobotTrajectoryPtr& o
     return;
   }
 
-  blend(other, blend_radius);
+  blend(planning_scene, other, blend_radius);
 }
 
 }  // namespace pilz_industrial_motion_planner
