@@ -40,13 +40,14 @@
 #include <tf2_kdl/tf2_kdl.h>
 #include <tf2_geometry_msgs/tf2_geometry_msgs.h>
 
-bool pilz_industrial_motion_planner::computePoseIK(const moveit::core::RobotModelConstPtr& robot_model,
+bool pilz_industrial_motion_planner::computePoseIK(const planning_scene::PlanningSceneConstPtr& scene,
                                                    const std::string& group_name, const std::string& link_name,
                                                    const Eigen::Isometry3d& pose, const std::string& frame_id,
                                                    const std::map<std::string, double>& seed,
                                                    std::map<std::string, double>& solution, bool check_self_collision,
                                                    const double timeout)
 {
+  const moveit::core::RobotModelConstPtr& robot_model = scene->getRobotModel();
   if (!robot_model->hasJointModelGroup(group_name))
   {
     ROS_ERROR_STREAM("Robot model has no planning group named as " << group_name);
@@ -74,7 +75,7 @@ bool pilz_industrial_motion_planner::computePoseIK(const moveit::core::RobotMode
 
   moveit::core::GroupStateValidityCallbackFn ik_constraint_function;
   ik_constraint_function =
-      boost::bind(&pilz_industrial_motion_planner::isStateColliding, check_self_collision, robot_model, _1, _2, _3);
+      boost::bind(&pilz_industrial_motion_planner::isStateColliding, check_self_collision, scene, _1, _2, _3);
 
   // call ik
   if (rstate.setFromIK(robot_model->getJointModelGroup(group_name), pose, link_name, timeout, ik_constraint_function))
@@ -93,7 +94,7 @@ bool pilz_industrial_motion_planner::computePoseIK(const moveit::core::RobotMode
   }
 }
 
-bool pilz_industrial_motion_planner::computePoseIK(const moveit::core::RobotModelConstPtr& robot_model,
+bool pilz_industrial_motion_planner::computePoseIK(const planning_scene::PlanningSceneConstPtr& scene,
                                                    const std::string& group_name, const std::string& link_name,
                                                    const geometry_msgs::Pose& pose, const std::string& frame_id,
                                                    const std::map<std::string, double>& seed,
@@ -102,7 +103,7 @@ bool pilz_industrial_motion_planner::computePoseIK(const moveit::core::RobotMode
 {
   Eigen::Isometry3d pose_eigen;
   tf2::fromMsg(pose, pose_eigen);
-  return computePoseIK(robot_model, group_name, link_name, pose_eigen, frame_id, seed, solution, check_self_collision,
+  return computePoseIK(scene, group_name, link_name, pose_eigen, frame_id, seed, solution, check_self_collision,
                        timeout);
 }
 
@@ -191,7 +192,7 @@ bool pilz_industrial_motion_planner::verifySampleJointLimits(
 }
 
 bool pilz_industrial_motion_planner::generateJointTrajectory(
-    const moveit::core::RobotModelConstPtr& robot_model,
+    const planning_scene::PlanningSceneConstPtr& scene,
     const pilz_industrial_motion_planner::JointLimitsContainer& joint_limits, const KDL::Trajectory& trajectory,
     const std::string& group_name, const std::string& link_name,
     const std::map<std::string, double>& initial_joint_position, const double& sampling_time,
@@ -200,6 +201,7 @@ bool pilz_industrial_motion_planner::generateJointTrajectory(
 {
   ROS_DEBUG("Generate joint trajectory from a Cartesian trajectory.");
 
+  const moveit::core::RobotModelConstPtr& robot_model = scene->getRobotModel();
   ros::Time generation_begin = ros::Time::now();
 
   // generate the time samples
@@ -225,7 +227,7 @@ bool pilz_industrial_motion_planner::generateJointTrajectory(
   {
     tf2::fromMsg(tf2::toMsg(trajectory.Pos(*time_iter)), pose_sample);
 
-    if (!computePoseIK(robot_model, group_name, link_name, pose_sample, robot_model->getModelFrame(), ik_solution_last,
+    if (!computePoseIK(scene, group_name, link_name, pose_sample, robot_model->getModelFrame(), ik_solution_last,
                        ik_solution, check_self_collision))
     {
       ROS_ERROR("Failed to compute inverse kinematics solution for sampled "
@@ -306,7 +308,7 @@ bool pilz_industrial_motion_planner::generateJointTrajectory(
 }
 
 bool pilz_industrial_motion_planner::generateJointTrajectory(
-    const moveit::core::RobotModelConstPtr& robot_model,
+    const planning_scene::PlanningSceneConstPtr& scene,
     const pilz_industrial_motion_planner::JointLimitsContainer& joint_limits,
     const pilz_industrial_motion_planner::CartesianTrajectory& trajectory, const std::string& group_name,
     const std::string& link_name, const std::map<std::string, double>& initial_joint_position,
@@ -315,6 +317,7 @@ bool pilz_industrial_motion_planner::generateJointTrajectory(
 {
   ROS_DEBUG("Generate joint trajectory from a Cartesian trajectory.");
 
+  const moveit::core::RobotModelConstPtr& robot_model = scene->getRobotModel();
   ros::Time generation_begin = ros::Time::now();
 
   std::map<std::string, double> ik_solution_last = initial_joint_position;
@@ -330,7 +333,7 @@ bool pilz_industrial_motion_planner::generateJointTrajectory(
   for (size_t i = 0; i < trajectory.points.size(); ++i)
   {
     // compute inverse kinematics
-    if (!computePoseIK(robot_model, group_name, link_name, trajectory.points.at(i).pose, robot_model->getModelFrame(),
+    if (!computePoseIK(scene, group_name, link_name, trajectory.points.at(i).pose, robot_model->getModelFrame(),
                        ik_solution_last, ik_solution, check_self_collision))
     {
       ROS_ERROR("Failed to compute inverse kinematics solution for sampled "
@@ -556,7 +559,7 @@ bool pilz_industrial_motion_planner::intersectionFound(const Eigen::Vector3d& p_
 }
 
 bool pilz_industrial_motion_planner::isStateColliding(const bool test_for_self_collision,
-                                                      const moveit::core::RobotModelConstPtr& robot_model,
+                                                      const planning_scene::PlanningSceneConstPtr& scene,
                                                       robot_state::RobotState* rstate,
                                                       const robot_state::JointModelGroup* const group,
                                                       const double* const ik_solution)
@@ -570,9 +573,9 @@ bool pilz_industrial_motion_planner::isStateColliding(const bool test_for_self_c
   rstate->update();
   collision_detection::CollisionRequest collision_req;
   collision_req.group_name = group->getName();
+  collision_req.verbose = true;
   collision_detection::CollisionResult collision_res;
-  planning_scene::PlanningScene(robot_model).checkSelfCollision(collision_req, collision_res, *rstate);
-
+  scene->checkSelfCollision(collision_req, collision_res, *rstate);
   return !collision_res.collision;
 }
 
@@ -581,4 +584,27 @@ void normalizeQuaternion(geometry_msgs::Quaternion& quat)
   tf2::Quaternion q;
   tf2::fromMsg(quat, q);
   quat = tf2::toMsg(q.normalize());
+}
+
+Eigen::Isometry3d getConstraintPose(const geometry_msgs::Point& position, const geometry_msgs::Quaternion& orientation,
+                                    const geometry_msgs::Vector3& offset)
+{
+  Eigen::Quaterniond quat;
+  tf2::fromMsg(orientation, quat);
+  quat.normalize();
+  Eigen::Vector3d v;
+  tf2::fromMsg(position, v);
+
+  Eigen::Isometry3d pose = Eigen::Translation3d(v) * quat;
+
+  tf2::fromMsg(offset, v);
+  pose.translation() -= quat * v;
+  return pose;
+}
+
+Eigen::Isometry3d getConstraintPose(const moveit_msgs::Constraints& goal)
+{
+  return getConstraintPose(goal.position_constraints.front().constraint_region.primitive_poses.front().position,
+                           goal.orientation_constraints.front().orientation,
+                           goal.position_constraints.front().target_point_offset);
 }
