@@ -51,7 +51,8 @@
 namespace pilz_industrial_motion_planner
 {
 TrajectoryGeneratorCIRC::TrajectoryGeneratorCIRC(const moveit::core::RobotModelConstPtr& robot_model,
-                                                 const LimitsContainer& planner_limits)
+                                                 const LimitsContainer& planner_limits,
+                                                 const std::string& /*group_name*/)
   : TrajectoryGenerator::TrajectoryGenerator(robot_model, planner_limits)
 {
   if (!planner_limits_.hasFullCartesianLimits())
@@ -84,7 +85,8 @@ void TrajectoryGeneratorCIRC::cmdSpecificRequestValidation(const planning_interf
   }
 }
 
-void TrajectoryGeneratorCIRC::extractMotionPlanInfo(const planning_interface::MotionPlanRequest& req,
+void TrajectoryGeneratorCIRC::extractMotionPlanInfo(const planning_scene::PlanningSceneConstPtr& scene,
+                                                    const planning_interface::MotionPlanRequest& req,
                                                     TrajectoryGenerator::MotionPlanInfo& info) const
 {
   ROS_DEBUG("Extract necessary information from motion plan request.");
@@ -140,12 +142,7 @@ void TrajectoryGeneratorCIRC::extractMotionPlanInfo(const planning_interface::Mo
     {
       frame_id = req.goal_constraints.front().position_constraints.front().header.frame_id;
     }
-    geometry_msgs::Pose goal_pose_msg;
-    goal_pose_msg.position =
-        req.goal_constraints.front().position_constraints.front().constraint_region.primitive_poses.front().position;
-    goal_pose_msg.orientation = req.goal_constraints.front().orientation_constraints.front().orientation;
-    normalizeQuaternion(goal_pose_msg.orientation);
-    tf2::fromMsg(goal_pose_msg, info.goal_pose);
+    info.goal_pose = getConstraintPose(req.goal_constraints.front());
   }
 
   assert(req.start_state.joint_state.name.size() == req.start_state.joint_state.position.size());
@@ -167,7 +164,7 @@ void TrajectoryGeneratorCIRC::extractMotionPlanInfo(const planning_interface::Mo
 
   // check goal pose ik before Cartesian motion plan starts
   std::map<std::string, double> ik_solution;
-  if (!computePoseIK(robot_model_, info.group_name, info.link_name, info.goal_pose, frame_id, info.start_joint_position,
+  if (!computePoseIK(scene, info.group_name, info.link_name, info.goal_pose, frame_id, info.start_joint_position,
                      ik_solution))
   {
     // LCOV_EXCL_START
@@ -177,16 +174,27 @@ void TrajectoryGeneratorCIRC::extractMotionPlanInfo(const planning_interface::Mo
     // LCOV_EXCL_STOP // not able to trigger here since lots of checks before
     // are in place
   }
-
-  Eigen::Vector3d circ_path_point;
-  tf2::fromMsg(req.path_constraints.position_constraints.front().constraint_region.primitive_poses.front().position,
-               circ_path_point);
-
   info.circ_path_point.first = req.path_constraints.name;
-  info.circ_path_point.second = circ_path_point;
+  if (!req.goal_constraints.front().position_constraints.empty())
+  {
+    const moveit_msgs::Constraints& goal = req.goal_constraints.front();
+    info.circ_path_point.second =
+        getConstraintPose(
+            req.path_constraints.position_constraints.front().constraint_region.primitive_poses.front().position,
+            goal.orientation_constraints.front().orientation, goal.position_constraints.front().target_point_offset)
+            .translation();
+  }
+  else
+  {
+    Eigen::Vector3d circ_path_point;
+    tf2::fromMsg(req.path_constraints.position_constraints.front().constraint_region.primitive_poses.front().position,
+                 circ_path_point);
+    info.circ_path_point.second = circ_path_point;
+  }
 }
 
-void TrajectoryGeneratorCIRC::plan(const planning_interface::MotionPlanRequest& req, const MotionPlanInfo& plan_info,
+void TrajectoryGeneratorCIRC::plan(const planning_scene::PlanningSceneConstPtr& scene,
+                                   const planning_interface::MotionPlanRequest& req, const MotionPlanInfo& plan_info,
                                    const double& sampling_time, trajectory_msgs::JointTrajectory& joint_trajectory)
 {
   std::unique_ptr<KDL::Path> cart_path(setPathCIRC(plan_info));
@@ -202,9 +210,9 @@ void TrajectoryGeneratorCIRC::plan(const planning_interface::MotionPlanRequest& 
   moveit_msgs::MoveItErrorCodes error_code;
   // sample the Cartesian trajectory and compute joint trajectory using inverse
   // kinematics
-  if (!generateJointTrajectory(robot_model_, planner_limits_.getJointLimitContainer(), cart_trajectory,
-                               plan_info.group_name, plan_info.link_name, plan_info.start_joint_position, sampling_time,
-                               joint_trajectory, error_code))
+  if (!generateJointTrajectory(scene, planner_limits_.getJointLimitContainer(), cart_trajectory, plan_info.group_name,
+                               plan_info.link_name, plan_info.start_joint_position, sampling_time, joint_trajectory,
+                               error_code))
   {
     throw CircTrajectoryConversionFailure("Failed to generate valid joint trajectory from the Cartesian path",
                                           error_code.val);
