@@ -119,7 +119,8 @@ protected:
     planner_limits.setCartesianLimits(cart_limits);
 
     // create planner instance
-    trajectory_generator_ = std::unique_ptr<typename T::Type_>(new typename T::Type_(robot_model_, planner_limits));
+    trajectory_generator_ =
+        std::unique_ptr<typename T::Type_>(new typename T::Type_(robot_model_, planner_limits, planning_group_));
     ASSERT_NE(nullptr, trajectory_generator_) << "failed to create trajectory generator";
 
     // create a valid motion plan request with goal in joint space as basis for
@@ -147,6 +148,7 @@ protected:
     robot_model_loader::RobotModelLoader(!T::VALUE ? PARAM_MODEL_NO_GRIPPER_NAME : PARAM_MODEL_WITH_GRIPPER_NAME)
         .getModel()
   };
+  planning_scene::PlanningSceneConstPtr planning_scene_{ new planning_scene::PlanningScene(robot_model_) };
 
   // trajectory generator
   std::unique_ptr<pilz_industrial_motion_planner::TrajectoryGenerator> trajectory_generator_;
@@ -177,22 +179,22 @@ TYPED_TEST_SUITE(TrajectoryGeneratorCommonTestWithGripper, TrajectoryGeneratorCo
 TYPED_TEST(TrajectoryGeneratorCommonTest, InvalideScalingFactor)
 {
   this->req_.max_velocity_scaling_factor = 2.0;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_MOTION_PLAN);
 
   this->req_.max_velocity_scaling_factor = 1.0;
   this->req_.max_acceleration_scaling_factor = 0;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_MOTION_PLAN);
 
   this->req_.max_velocity_scaling_factor = 0.00001;
   this->req_.max_acceleration_scaling_factor = 1;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_MOTION_PLAN);
 
   this->req_.max_velocity_scaling_factor = 1;
   this->req_.max_acceleration_scaling_factor = -1;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_MOTION_PLAN);
 }
 
@@ -202,7 +204,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, InvalideScalingFactor)
 TYPED_TEST(TrajectoryGeneratorCommonTest, InvalidGroupName)
 {
   this->req_.group_name = "foot";
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(moveit_msgs::MoveItErrorCodes::INVALID_GROUP_NAME, this->res_.error_code_.val);
 }
 
@@ -212,7 +214,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, InvalidGroupName)
 TYPED_TEST(TrajectoryGeneratorCommonTestNoGripper, GripperGroup)
 {
   this->req_.group_name = "gripper";
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(moveit_msgs::MoveItErrorCodes::INVALID_GROUP_NAME, this->res_.error_code_.val);
 }
 
@@ -222,7 +224,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTestNoGripper, GripperGroup)
 TYPED_TEST(TrajectoryGeneratorCommonTestWithGripper, GripperGroup)
 {
   this->req_.group_name = "gripper";
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS, this->res_.error_code_.val);
 }
 
@@ -247,7 +249,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTestWithGripper, GripperGroup)
 TYPED_TEST(TrajectoryGeneratorCommonTest, EmptyJointNamesInStartState)
 {
   this->req_.start_state.joint_state.name.clear();
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_ROBOT_STATE);
 }
 
@@ -257,7 +259,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, EmptyJointNamesInStartState)
 TYPED_TEST(TrajectoryGeneratorCommonTest, InconsistentStartState)
 {
   this->req_.start_state.joint_state.name.push_back("joint_7");
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_ROBOT_STATE);
 }
 
@@ -267,7 +269,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, InconsistentStartState)
 TYPED_TEST(TrajectoryGeneratorCommonTest, StartPostionOutOfLimit)
 {
   this->req_.start_state.joint_state.position[0] = 100;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_ROBOT_STATE);
 }
 
@@ -281,7 +283,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, StartPostionOutOfLimit)
 TYPED_TEST(TrajectoryGeneratorCommonTest, StartPositionVelocityNoneZero)
 {
   this->req_.start_state.joint_state.velocity[0] = 100;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_ROBOT_STATE);
 }
 
@@ -291,7 +293,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, StartPositionVelocityNoneZero)
 TYPED_TEST(TrajectoryGeneratorCommonTest, EmptyGoalConstraints)
 {
   this->req_.goal_constraints.clear();
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
@@ -308,7 +310,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, MultipleGoals)
   // two goal constraints
   this->req_.goal_constraints.push_back(goal_constraint);
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 
   // one joint constraint and one orientation constraint
@@ -316,7 +318,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, MultipleGoals)
   goal_constraint.orientation_constraints.push_back(orientation_constraint);
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 
   // one joint constraint and one Cartesian constraint
@@ -324,7 +326,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, MultipleGoals)
   goal_constraint.orientation_constraints.push_back(orientation_constraint);
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 
   // two Cartesian constraints
@@ -335,7 +337,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, MultipleGoals)
   goal_constraint.orientation_constraints.push_back(orientation_constraint);
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
@@ -347,7 +349,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, InvalideJointNameInGoal)
   moveit_msgs::JointConstraint joint_constraint;
   joint_constraint.joint_name = "test_joint_2";
   this->req_.goal_constraints.front().joint_constraints[0] = joint_constraint;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
@@ -359,7 +361,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, MissingJointConstraint)
   moveit_msgs::JointConstraint joint_constraint;
   joint_constraint.joint_name = "test_joint_2";
   this->req_.goal_constraints.front().joint_constraints.pop_back();  //<-- Missing joint constraint
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
@@ -369,7 +371,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, MissingJointConstraint)
 TYPED_TEST(TrajectoryGeneratorCommonTest, InvalideJointPositionInGoal)
 {
   this->req_.goal_constraints.front().joint_constraints[0].position = 100;
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
@@ -386,7 +388,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, InvalidLinkNameInCartesianGoal)
   goal_constraint.orientation_constraints.push_back(orientation_constraint);
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 
   // different link names in position and orientation goals
@@ -394,14 +396,14 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, InvalidLinkNameInCartesianGoal)
   goal_constraint.orientation_constraints.front().link_name = "test_link_2";
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 
   // no solver for the link
   goal_constraint.orientation_constraints.front().link_name = "test_link_1";
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::NO_IK_SOLUTION);
 }
 
@@ -421,7 +423,7 @@ TYPED_TEST(TrajectoryGeneratorCommonTest, EmptyPrimitivePoses)
   goal_constraint.orientation_constraints.push_back(orientation_constraint);
   this->req_.goal_constraints.clear();
   this->req_.goal_constraints.push_back(goal_constraint);
-  EXPECT_FALSE(this->trajectory_generator_->generate(this->req_, this->res_));
+  EXPECT_FALSE(this->trajectory_generator_->generate(this->planning_scene_, this->req_, this->res_));
   EXPECT_EQ(this->res_.error_code_.val, moveit_msgs::MoveItErrorCodes::INVALID_GOAL_CONSTRAINTS);
 }
 
